Endpoints accept `annotations` param
