+++ app/controllers/admin/api/backend_apis_controller.rb
@@ -8,7 +8,7 @@ class Admin::Api::BackendApisController < Admin::Api::BaseController
   clear_respond_to
   respond_to :json
 
-  wrap_parameters BackendApi, include: BackendApi.attribute_names | %w[managed_by]
+  wrap_parameters BackendApi, include: BackendApi.attribute_names | %w[annotations]
   representer BackendApi
 
   paginate only: :index
@@ -48,7 +48,7 @@ class Admin::Api::BackendApisController < Admin::Api::BaseController
 
   private
 
-  DEFAULT_PARAMS = %i[name description private_endpoint managed_by].freeze
+  DEFAULT_PARAMS = %i[name description private_endpoint annotations].freeze
   private_constant :DEFAULT_PARAMS
 
   def authorize
