+++ app/controllers/master/api/providers_controller.rb
@@ -82,7 +82,7 @@ class Master::Api::ProvidersController < Master::Api::BaseController
     end
   end
 
-  UPDATE_PARAMS = %i[from_email support_email finance_support_email site_access_code state_event managed_by].freeze
+  UPDATE_PARAMS = [:from_email, :support_email, :finance_support_email, :site_access_code, :state_event, {annotations: %i[name value]}].freeze
   private_constant :UPDATE_PARAMS
 
   private
