+++ app/representers/account_representer.rb
@@ -51,7 +51,6 @@ module AccountRepresenter
       property :credit_card_partial_number
       property :credit_card_expires_on
     end
-    property :managed_by
   end
 
   property :state
