+++ app/representers/backend_api_representer.rb
@@ -10,7 +10,6 @@ module BackendApiRepresenter
   property :system_name
   property :description
   property :private_endpoint
-  property :managed_by
   property :account_id
   property :created_at
   property :updated_at
